+++ tests/test_sklearn_compat.py
@@ -15,7 +15,6 @@ from scikit_opls import OPLS, OPLSDA
         OPLS(n_orthogonal=0),
         OPLS(scale="pareto"),
         OPLSDA(),
-        OPLSDA(probability=True),
     ]
 )
 def test_sklearn_compatible_estimator(estimator, check):